Move deploy function from test_charm.py to conftest.py
