+++ postfix-relay-operator/tests/integration/conftest.py
@@ -11,6 +11,23 @@ import pytest
 APP_NAME = "postfix-relay"
 
 
+def deploy(juju: jubilant.Juju, charm: str) -> None:
+    """Deploy postfix-relay and its dependencies.
+
+    Args:
+        juju: Jubilant Juju instance.
+        charm: Path to the charm file to deploy.
+    """
+    juju.deploy(f"./{charm}", APP_NAME)
+    juju.deploy("self-signed-certificates", channel="latest/edge")
+    juju.integrate(APP_NAME, "self-signed-certificates")
+    juju.wait(
+        lambda status: status.apps[APP_NAME].is_active,
+        error=jubilant.any_blocked,
+        timeout=10 * 60,
+    )
+
+
 @pytest.fixture(scope="module")
 def juju(juju: jubilant.Juju) -> jubilant.Juju:
     """Override juju fixture to set wait timeout."""
